change lr

+++ vqvae/quantizers.py
@@ -57,7 +57,7 @@ class VectorQuantizer(tf.keras.layers.Layer):
 		if self.is_training:
 			# Percentage of unconditional training
 			uncondition_mask = tf.random.uniform((input_shape[0], ))
-			random_indices = tf.random.uniform(tf.shape(encoding_indices), minval = 0, maxval = self.num_embeddings, dtype = tf.int32)
+			random_indices = tf.random.uniform(tf.shape(encoding_indices), minval = 0, maxval = self.num_embeddings, dtype = tf.int64)
 			encoding_indices = tf.where(
 				uncondition_mask > (UNCONDITION_PROB * 2),
 				encoding_indices,
